Güvenli Sınav Denetim Sistemi/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
pymysql==1.1.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
websockets==12.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.4
